backend/pyproject.toml
[project]
name = "pdftools-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-multipart>=0.0.12",
    "pypdf>=5.1",
    "PyMuPDF>=1.24",
    "Pillow>=11.0",
    "numpy>=2.1",
    "cryptography>=43.0",
    "apscheduler>=3.10",
    "pydantic-settings>=2.6",
    "sqlmodel>=0.0.22",
    "img2pdf>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
